api/requirements.txt
Flask
cryptography==1.2.3
mysql-python
Flask-SQLAlchemy
SQLAlchemy
gunicorn
requests
voluptuous
Pillow
markdown2
onetimepass
pyqrcode
python-memcached
coverage
gitpython
pycrypto
pyyaml
paramiko